+++ theme/gatsby-node.spec.js
@@ -90,6 +90,7 @@ describe('gatsby-node', () => {
       expect(typeDefs).toContain('slug: String!')
       expect(typeDefs).toContain('text: String!')
       expect(typeDefs).toContain('title: String!')
+      expect(typeDefs).toContain('nativeAnchor: Boolean')
 
       // Check that widget config has optional fields
       expect(typeDefs).toContain('username: String')